requirements.txt
asposestorage==1.0.2
backpack==0.1
backpack_for_pytorch==1.5.0
matplotlib==3.5.1
numpy==1.21.2
Pillow==9.1.1
prv_accountant==0.1.1.post1
requests==2.27.1
scikit_learn==1.1.1
scipy==1.8.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
